# docs/swagger.yaml
basePath: /
definitions:
...
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      deleted_at:
        format: date-time
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  handler.Meta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  handler.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  handler.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  handler.UpdateUserResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/handler.UserResponse'
    type: object
  handler.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      meta:
        $ref: '#/definitions/handler.Meta'
    type: object
  handler.UserResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
host: localhost:7777
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Elysian Backend API provides user authentication, management, and health
    check endpoints. Built with Go and Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Elysian Backend API
  version: 1.0.0
paths:
...
  /api/v1/users:
    get:
      description: Get list of users
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List users
      tags:
      - users
  /api/v1/users/{id}:
    get:
      description: Get user details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
  /api/v1/users/email/{email}:
    get:
      description: Get user details by email
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get user by email
      tags:
      - users
  /api/v1/users/me:
    delete:
      description: Delete currently logged in user account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - users
    get:
      description: Get details of currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update details of currently logged in user
      parameters:
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"