# internal/delivery/http/handler/user_handler.go
package handler

import (
	"encoding/json"
	"net/http"
	"strconv"
	"time"

	"github.com/Elysian-Rebirth/backend-go/internal/domain"
	"github.com/Elysian-Rebirth/backend-go/internal/domain/repository"
	"github.com/Elysian-Rebirth/backend-go/internal/middleware"
	"github.com/gin-gonic/gin"
	"golang.org/x/crypto/bcrypt"
	"gorm.io/datatypes"
)

type UserHandler struct {
	userRepo repository.UserRepository
}

func NewUserHandler(userRepo repository.UserRepository) *UserHandler {
	return &UserHandler{
		userRepo: userRepo,
	}
}

// Request and Response structs
type UpdateUserRequest struct {
	Name      string   `json:"name" validate:"min=2,max=100"`
	AvatarURL *string  `json:"avatar_url"`
	Bio       *string  `json:"bio"`
	Links     []string `json:"links"`
}

type UserResponse struct {
	ID        string    `json:"id"`
	Email     string    `json:"email"`
	Name      string    `json:"name"`
	AvatarURL *string   `json:"avatar_url,omitempty"`
	Bio       string    `json:"bio"`
	Links     []string  `json:"links"`
	IsActive  bool      `json:"is_active"`
	CreatedAt time.Time `json:"created_at"`
}

type UserListResponse struct {
	Data []*domain.User `json:"data"`
	Meta Meta           `json:"meta"`
}

type Meta struct {
	Total  int64 `json:"total"`
	Limit  int   `json:"limit"`
	Offset int   `json:"offset"`
}

type UpdateUserResponse struct {
	Message string       `json:"message"`
	User    UserResponse `json:"user"`
}

// GetByID godoc
// @Summary      Get user by ID
// @Description  Get user details by ID
// @Tags         users
// @Produce      json
// @Param        id   path      string  true  "User ID"
// @Success      200  {object}  domain.User
// @Failure      404  {object}  ErrorResponse
// @Router       /api/v1/users/{id} [get]
func (h *UserHandler) GetByID(c *gin.Context) {
	id := c.Param("id")

	user, err := h.userRepo.FindByID(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{Error: "User not found"})
		return
	}

	c.JSON(http.StatusOK, user)
}

// List godoc
// @Summary      List users
// @Description  Get list of users
// @Tags         users
// @Produce      json
// @Param        limit   query     int     false  "Limit"
// @Param        offset  query     int     false  "Offset"
// @Success      200     {object}  UserListResponse
// @Failure      500     {object}  ErrorResponse
// @Router       /api/v1/users [get]
func (h *UserHandler) List(c *gin.Context) {
	limitStr := c.DefaultQuery("limit", "10")
	offsetStr := c.DefaultQuery("offset", "0")

	limit, err := strconv.Atoi(limitStr)
	if err != nil || limit < 1 || limit > 100 {
		limit = 10
	}

	offset, err := strconv.Atoi(offsetStr)
	if err != nil || offset < 0 {
		offset = 0
	}

	users, total, err := h.userRepo.List(c.Request.Context(), limit, offset)
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to fetch users"})
		return
	}

	c.JSON(http.StatusOK, UserListResponse{
		Data: users,
		Meta: Meta{
			Total:  total,
			Limit:  limit,
			Offset: offset,
		},
	})
}

// GetByEmail godoc
// @Summary      Get user by email
// @Description  Get user details by email
// @Tags         users
// @Produce      json
// @Param        email path      string  true  "User Email"
// @Success      200   {object}  domain.User
// @Failure      404   {object}  ErrorResponse
// @Router       /api/v1/users/email/{email} [get]
func (h *UserHandler) GetByEmail(c *gin.Context) {
	email := c.Param("email")

	user, err := h.userRepo.FindByEmail(c.Request.Context(), email)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{Error: "User not found"})
		return
	}

	c.JSON(http.StatusOK, user)
}

// GetMe godoc
// @Summary      Get current user
// @Description  Get details of currently logged in user
// @Tags         users
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  UserResponse
// @Router       /api/v1/users/me [get]
func (h *UserHandler) GetMe(c *gin.Context) {
	currentUser := middleware.MustGetUserFromContext(c)

	// Fetch fresh user data
	user, err := h.userRepo.FindByID(c.Request.Context(), currentUser.ID.String())
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to fetch user profile"})
		return
	}

	// Fetch extended profile (bio, links)
	profile, err := h.userRepo.GetProfile(c.Request.Context(), user.ID.String())
	if err != nil {
		profile = &domain.UserProfile{Bio: "", LinksJSON: []byte("[]")}
	}

	// Decode links JSON
	var links []string
	if err := json.Unmarshal(profile.LinksJSON, &links); err != nil {
		links = []string{}
	}

	c.JSON(http.StatusOK, gin.H{
		"status": "success",
		"data": gin.H{
			"id":         user.ID.String(),
			"email":      user.Email,
			"full_name":  user.FullName,
			"avatar_url": user.AvatarURL,
			"bio":        profile.Bio,
			"links":      links,
			"created_at": user.CreatedAt,
			"updated_at": user.UpdatedAt,
		},
	})
}

// UpdateMe godoc
// @Summary      Update current user
// @Description  Update details of currently logged in user
// @Tags         users
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request body UpdateUserRequest true "Update Request"
// @Success      200  {object}  UpdateUserResponse
// @Failure      400  {object}  ErrorResponse
// @Failure      500  {object}  ErrorResponse
// @Router       /api/v1/users/me [put]
func (h *UserHandler) UpdateMe(c *gin.Context) {
	user := middleware.MustGetUserFromContext(c)

	var req UpdateUserRequest

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: "Invalid request body"})
		return
	}

	// Update core user fields
	if req.Name != "" {
		user.FullName = req.Name
	}
	if req.AvatarURL != nil {
		user.AvatarURL = *req.AvatarURL
	}

	if err := h.userRepo.Update(c.Request.Context(), user); err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to update profile"})
		return
	}

	// Upsert extended profile (bio + links)
	uid := user.ID
	profile := &domain.UserProfile{
		UserID: uid,
		Bio:    "",
	}
	if req.Bio != nil {
		profile.Bio = *req.Bio
	}

	// Encode links to JSON
	links := req.Links
	if links == nil {
		links = []string{}
	}
	linksJSON, err := json.Marshal(links)
	if err != nil {
		linksJSON = []byte("[]")
	}
	profile.LinksJSON = datatypes.JSON(linksJSON)

	if err := h.userRepo.UpdateProfile(c.Request.Context(), profile); err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to update extended profile"})
		return
	}

	c.JSON(http.StatusOK, UpdateUserResponse{
		Message: "Profile updated successfully",
		User: UserResponse{
			ID:        user.ID.String(),
			Email:     user.Email,
			Name:      user.FullName,
			AvatarURL: &user.AvatarURL,
			Bio:       profile.Bio,
			Links:     links,
		},
	})
}

// DeleteMe godoc
// @Summary      Delete current user
// @Description  Delete currently logged in user account
// @Tags         users
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  SuccessResponse
// @Failure      500  {object}  ErrorResponse
// @Router       /api/v1/users/me [delete]
func (h *UserHandler) DeleteMe(c *gin.Context) {
	user := middleware.MustGetUserFromContext(c)

	if err := h.userRepo.Delete(c.Request.Context(), user.ID.String()); err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to delete account"})
		return
	}

	c.JSON(http.StatusOK, SuccessResponse{
		Message: "Account deleted successfully",
	})
}

type UpdatePasswordRequest struct {
	CurrentPassword string `json:"current_password" binding:"required"`
	NewPassword     string `json:"new_password" binding:"required" validate:"min=8"`
}

func (h *UserHandler) UpdatePassword(c *gin.Context) {
	user := middleware.MustGetUserFromContext(c)

	var req UpdatePasswordRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: err.Error()})
		return
	}

	freshUser, err := h.userRepo.FindByID(c.Request.Context(), user.ID.String())
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to fetch user profile"})
		return
	}

	err = bcrypt.CompareHashAndPassword([]byte(freshUser.PasswordHash), []byte(req.CurrentPassword))
	if err != nil {
		c.JSON(http.StatusUnauthorized, ErrorResponse{Error: "Incorrect current password"})
		return
	}

	newHash, err := bcrypt.GenerateFromPassword([]byte(req.NewPassword), bcrypt.DefaultCost)
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to hash new password"})
		return
	}

	freshUser.PasswordHash = string(newHash)
	if err := h.userRepo.Update(c.Request.Context(), freshUser); err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: "Failed to update password"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"status": "success", "message": "Password updated successfully"})
}

func (h *UserHandler) GetPreferences(c *gin.Context) {
	user := middleware.MustGetUserFromContext(c)

	prefs, err := h.userRepo.GetPreferences(c.Request.Context(), user.ID.String())
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"status": "success", "data": prefs})
}

type UpdatePreferencesRequest struct {
	Appearance       string         `json:"appearance"`
	Notifications    datatypes.JSON `json:"notifications"`
	SecuritySettings datatypes.JSON `json:"security_settings"`
}

func (h *UserHandler) UpdatePreferences(c *gin.Context) {
	user := middleware.MustGetUserFromContext(c)

	var req UpdatePreferencesRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{Error: err.Error()})
		return
	}

	prefs, err := h.userRepo.GetPreferences(c.Request.Context(), user.ID.String())
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: err.Error()})
		return
	}

	if req.Appearance != "" {
		prefs.Appearance = req.Appearance
	}
	if req.Notifications != nil {
		prefs.NotificationsJSON = req.Notifications
	}
	if req.SecuritySettings != nil {
		prefs.SecuritySettingsJSON = req.SecuritySettings
	}

	if err := h.userRepo.UpdatePreferences(c.Request.Context(), prefs); err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{Error: err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"status": "success", "data": prefs})
}